refactor: split near-cap orders lib modules (#4317 Slice B)
Extract feedstock helpers, probe validator/ICE builder, and order-engine
full-pass validation run so industry_counsel_ranking, order_suggestion_probe,
and order_engine each have ≥30 lines headroom under the 350-line ceiling.

Refs #4317
