+++ packages/colonizethis_orders/lib/src/orders/order_suggestion_probe.dart
@@ -4,126 +4,8 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_diplomacy/colonizethis_diplomacy.dart';
 import 'incremental_candidate_validator.dart';
 import 'order_resolution_context.dart';
-
-bool _orderSuggestionTrackWorkOrderAcceptanceProbes = false;
-int _orderSuggestionWorkOrderAcceptanceProbeCount = 0;
-
-/// Test hook: enable counting of order-engine work-order acceptance probes.
-void setOrderSuggestionWorkOrderAcceptanceProbeTrackingForTests(bool enabled) {
-  _orderSuggestionTrackWorkOrderAcceptanceProbes = enabled;
-  _orderSuggestionWorkOrderAcceptanceProbeCount = 0;
-}
-
-/// Test hook: probes counted while tracking is enabled (Refs #2133).
-int get orderSuggestionWorkOrderAcceptanceProbeCountForTests =>
-    _orderSuggestionWorkOrderAcceptanceProbeCount;
-
-void bumpOrderSuggestionWorkOrderAcceptanceProbeIfTracking() {
-  if (_orderSuggestionTrackWorkOrderAcceptanceProbes) {
-    _orderSuggestionWorkOrderAcceptanceProbeCount++;
-  }
-}
-
-int _incrementalCandidateValidatorBuildCountForTests = 0;
-
-/// Test hook: reset [incrementalCandidateValidatorBuildCountForTests] (Refs #2394).
-void resetIncrementalCandidateValidatorBuildCountForTests() {
-  _incrementalCandidateValidatorBuildCountForTests = 0;
-}
-
-/// Test hook: [buildIncrementalCandidateValidator] invocations since last reset.
-int get incrementalCandidateValidatorBuildCountForTests =>
-    _incrementalCandidateValidatorBuildCountForTests;
-
-IncrementalCandidateValidator buildIncrementalCandidateValidator({
-  required Game game,
-  required MapTopology topology,
-  required String playerId,
-  required Orders baseOrders,
-  Map<String, TileMapResult>? tileMapByRegion,
-  OrderResolutionContext? resolution,
-
-  /// When callers already built membership for this [game], pass it to avoid a
-  /// second [DiplomacyFactionMembership.from] inside the validator.
-  DiplomacyFactionMembership? factionMembership,
-}) {
-  _incrementalCandidateValidatorBuildCountForTests++;
-  return IncrementalCandidateValidator.forPlayer(
-    game: game,
-    topology: topology,
-    playerId: playerId,
-    basePrefix: baseOrders,
-    tileMapByRegion: tileMapByRegion,
-    resolution: resolution,
-    factionMembership: factionMembership,
-  );
-}
-
-/// Resolves the [IncrementalCandidateValidator] for a stateless candidate probe.
-///
-/// When [sharedCandidateValidator] is supplied for the same suggestion pass,
-/// rebinds via [IncrementalCandidateValidator.forBasePrefix] when [baseOrders]
-/// differs from the embedded prefix; otherwise reuses the instance without
-/// incrementing [incrementalCandidateValidatorBuildCountForTests]. Refs #2394.
-IncrementalCandidateValidator incrementalValidatorForCandidateProbe({
-  required Game game,
-  required MapTopology topology,
-  required String playerId,
-  required Orders baseOrders,
-  Map<String, TileMapResult>? tileMapByRegion,
-  OrderResolutionContext? resolution,
-  DiplomacyFactionMembership? factionMembership,
-  IncrementalCandidateValidator? sharedCandidateValidator,
-}) {
-  assert(
-    sharedCandidateValidator == null ||
-        sharedCandidateValidator.playerId == playerId,
-    'sharedCandidateValidator playerId must match playerId',
-  );
-  final shared = sharedCandidateValidator;
-  if (shared != null) {
-    return shared.basePrefix == baseOrders
-        ? shared
-        : shared.forBasePrefix(baseOrders);
-  }
-  return buildIncrementalCandidateValidator(
-    game: game,
-    topology: topology,
-    playerId: playerId,
-    baseOrders: baseOrders,
-    tileMapByRegion: tileMapByRegion,
-    resolution: resolution,
-    factionMembership: factionMembership,
-  );
-}
-
-bool _probeOrderAccepted<T>({
-  required Game game,
-  required MapTopology topology,
-  required String playerId,
-  required Orders baseOrders,
-  required T candidate,
-  required bool Function(IncrementalCandidateValidator validator, T candidate)
-  probeWithValidator,
-  Map<String, TileMapResult>? tileMapByRegion,
-  IncrementalCandidateValidator? sharedCandidateValidator,
-  OrderResolutionContext? resolution,
-  DiplomacyFactionMembership? factionMembership,
-  void Function()? onBeforeProbe,
-}) {
-  final validator = incrementalValidatorForCandidateProbe(
-    game: game,
-    topology: topology,
-    playerId: playerId,
-    baseOrders: baseOrders,
-    tileMapByRegion: tileMapByRegion,
-    resolution: resolution,
-    factionMembership: factionMembership,
-    sharedCandidateValidator: sharedCandidateValidator,
-  );
-  onBeforeProbe?.call();
-  return probeWithValidator(validator, candidate);
-}
+import 'order_suggestion_probe_validator.dart';
+export 'order_suggestion_probe_validator.dart';
 
 bool isMoveOrderAccepted(
   Game game,
@@ -140,7 +22,7 @@ bool isMoveOrderAccepted(
   // [validatePlayerOrdersWithContext]. SPEC/program/order-suggestions.md
   // § Incremental candidate validation; SPEC/program/order-engine.md
   // § Validation (candidate-probe context). Refs #2237.
-  return _probeOrderAccepted<MoveOrder>(
+  return probeOrderAccepted<MoveOrder>(
     game: game,
     topology: topology,
     playerId: playerId,
@@ -167,7 +49,7 @@ bool isArmyMoveOrderAccepted(
   // [baseOrders]'s diplomatic context without re-running full-pass
   // [validatePlayerOrdersWithContext]. SPEC/program/order-suggestions.md
   // § Incremental candidate validation. Refs #2237.
-  return _probeOrderAccepted<ArmyMoveOrder>(
+  return probeOrderAccepted<ArmyMoveOrder>(
     game: game,
     topology: topology,
     playerId: playerId,
@@ -191,7 +73,7 @@ bool isWorkOrderAccepted(
   OrderResolutionContext? resolution,
   DiplomacyFactionMembership? factionMembership,
 }) {
-  return _probeOrderAccepted<WorkOrder>(
+  return probeOrderAccepted<WorkOrder>(
     game: game,
     topology: topology,
     playerId: playerId,
@@ -216,7 +98,7 @@ bool isBuildOrderAccepted(
   OrderResolutionContext? resolution,
   DiplomacyFactionMembership? factionMembership,
 }) {
-  return _probeOrderAccepted<BuildUnitOrder>(
+  return probeOrderAccepted<BuildUnitOrder>(
     game: game,
     topology: topology,
     playerId: playerId,
@@ -241,7 +123,7 @@ bool isNavalMoveOrderAccepted(
 }) {
   // Stateless candidate-probe path. SPEC/program/order-suggestions.md
   // § Incremental candidate validation. Refs #2237.
-  return _probeOrderAccepted<NavalMoveOrder>(
+  return probeOrderAccepted<NavalMoveOrder>(
     game: game,
     topology: topology,
     playerId: playerId,
@@ -266,7 +148,7 @@ bool isNavalMissionOrderAccepted(
 }) {
   // Stateless candidate-probe path. SPEC/program/order-suggestions.md
   // § Incremental candidate validation. Refs #2237.
-  return _probeOrderAccepted<NavalMissionOrder>(
+  return probeOrderAccepted<NavalMissionOrder>(
     game: game,
     topology: topology,
     playerId: playerId,
@@ -296,7 +178,7 @@ bool isDiplomaticOrderAccepted(
   DiplomacyFactionMembership? factionMembership,
   IncrementalCandidateValidator? sharedCandidateValidator,
 }) {
-  return _probeOrderAccepted<DiplomaticOrder>(
+  return probeOrderAccepted<DiplomaticOrder>(
     game: game,
     topology: topology,
     playerId: playerId,
